+++ lib/mongoid/timestamps/created.rb
@@ -1,6 +1,5 @@
 # encoding: utf-8
 module Mongoid #:nodoc:
-
   module Timestamps
     # This module handles the behaviour for setting up document created at
     # timestamp.
@@ -9,13 +8,7 @@ module Mongoid #:nodoc:
 
       included do
         field :created_at, :type => Time
-
-        set_callback :create, :before, :set_created_at
-
-        unless methods.include? 'record_timestamps'
-          class_attribute :record_timestamps
-          self.record_timestamps = true
-        end
+        set_callback :create, :before, :set_created_at, :if => :timestamping?
       end
 
       # Update the created_at field on the Document to the current time. This is